+++ Mapping/Loader/AbstractLoader.php
@@ -94,10 +94,16 @@ abstract class AbstractLoader implements LoaderInterface
             }
 
             if (1 === \count($options) && isset($options['value'])) {
+                if (\func_num_args() < 3 || !func_get_arg(2)) {
+                    trigger_deprecation('symfony/validator', '7.4', 'Using the "value" option to configure the "%s" constraint is deprecated.', $className);
+                }
+
                 return new $className($options['value']);
             }
 
             if (array_is_list($options)) {
+                trigger_deprecation('symfony/validator', '7.4', 'Configuring the "%s" without passing its option names is deprecated.', $className);
+
                 return new $className($options);
             }
 
@@ -105,6 +111,8 @@ abstract class AbstractLoader implements LoaderInterface
                 return new $className(...$options);
             } catch (\Error $e) {
                 if (str_starts_with($e->getMessage(), 'Unknown named parameter ')) {
+                    trigger_deprecation('symfony/validator', '7.4', 'Using option names not matching the named arguments of the "%s" constraint is deprecated.', $className);
+
                     return new $className($options);
                 }